requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.13
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
openai>=1.24.0
tiktoken==0.6.0
qdrant-client
pymongo>=4.6,<4.10
motor==3.3.2
redis
streamlit
pandas